decouple node details from chainspec type
